split `get` into its own function
