+++ app/src/js/v/get/index.js
@@ -1,0 +1,3 @@
+import get from './get';
+
+export default get;
